[Draft] Support queries for folding
Use queries to define fold ranges rather than individual nodes.

Rather than defining a mapping of node types to the methods which handle them, we instead define a mapping of _query sexps_ to their handlers. Each handler takes the beginning and end positions of the block to be folded.

When the buffer is re-parsed by tree-sitter, we re-run the relevant queries and generate a mapping between each node in the matched query and the encompassing fold range. Then, when a fold query comes in, we walk up the tree until we find a node which has a fold mapping. This fold mapping's beginning and end is passed into the handler and then we fold as usual.

In a given fold query, we can define specific captures. By default, the `@fold` capture is the bit which is folded.

Fold queries are defined in both the `ts-fold--default-patterns` and `ts-fold--patterns-to-handlers-list` variables. The latter allows users to define their own fold patterns, keyed by language symbol.